Fixes #12906: Make third-party authentication a plugin
