+++ rudder-core/src/main/scala/com/normation/rudder/domain/eventlog/PromisesEventLog.scala
@@ -38,7 +38,6 @@
 package com.normation.rudder.domain.eventlog
 
 import com.normation.eventlog._
-import com.normation.rudder.batch.StartDeploymentMessage
 import com.normation.utils.HashcodeCaching
 
 sealed trait PromiseEventLog extends EventLog { override final val eventLogCategory = DeploymentLogCategory }
