--- rudder-web/src/main/resources/applicationContext-security-auth-ldap.xml
@@ -1,71 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--
-Copyright 2014 Normation SAS
-
-This file is part of Rudder.
-
-Rudder is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-In accordance with the terms of section 7 (7. Additional Terms.) of
-the GNU General Public License version 3, the copyright holders add
-the following Additional permissions:
-Notwithstanding to the terms of section 5 (5. Conveying Modified Source
-Versions) and 6 (6. Conveying Non-Source Forms.) of the GNU General
-Public License version 3, when you create a Related Module, this
-Related Module is not considered as a part of the work and may be
-distributed under the license agreement of your choice.
-A "Related Module" means a set of sources files including their
-documentation that, without modification of the Source Code, enables
-supplementary functions or services in addition to those offered by
-the Software.
-
-Rudder is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with Rudder.  If not, see <http://www.gnu.org/licenses/>.
-
--->
-
-<beans:beans xmlns="http://www.springframework.org/schema/security"
-    xmlns:beans="http://www.springframework.org/schema/beans"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.2.xsd
-                        http://www.springframework.org/schema/security http://www.springframework.org/schema/security/spring-security-3.2.xsd">
-
-    <!-- 
-      The properties in that file are read from rudder-web.properties, then translated into
-      system properties in LiftInitContextListener class. 
-    -->
-    
-    <beans:bean id="contextSource" class="org.springframework.security.ldap.DefaultSpringSecurityContextSource">
-      <beans:constructor-arg value="${rudder.auth.ldap.connection.url}"/>
-      <beans:property name="userDn" value="${rudder.auth.ldap.connection.bind.dn}"/>
-      <beans:property name="password" value="${rudder.auth.ldap.connection.bind.password}"/>
-    </beans:bean>
-    
-    
-    <beans:bean id="userLookup" class="org.springframework.security.ldap.search.FilterBasedLdapUserSearch">
-      <beans:constructor-arg index="0" value="${rudder.auth.ldap.searchbase}"/>
-      <beans:constructor-arg index="1" value="${rudder.auth.ldap.filter}"/>
-      <beans:constructor-arg index="2" ref="contextSource" />
-    </beans:bean>
-    
-    <beans:bean id="ldapAuthenticationProvider" class="org.springframework.security.ldap.authentication.LdapAuthenticationProvider">
-       <beans:constructor-arg>
-         <beans:bean class="org.springframework.security.ldap.authentication.BindAuthenticator">
-           <beans:constructor-arg ref="contextSource"/>
-           <beans:property name="userSearch" ref="userLookup"/>
-         </beans:bean>
-       </beans:constructor-arg>
-     <beans:property name="userDetailsContextMapper" ref="rudderXMLUserDetails"/>
-    </beans:bean>
-
-
-</beans:beans>
